+++ plugins/network/src/commonMain/kotlin/com/ae/log/network/config/InterceptorDefaults.kt
@@ -37,6 +37,7 @@ public object InterceptorDefaults {
             "Accept-Language",
             "Connection",
             "Content-Type",
+            "Content-Length",
             "Host",
             "User-Agent",
             // Noisy system response headers